Refact some infer-doc-type import lines
