+++ lib/schema.js
@@ -108,7 +108,7 @@ function Schema(obj, options) {
   this._indexes = [];
   this.methods = {};
   this.methodOptions = {};
-  this.statics = options?.statics || {};
+  this.statics = (options && options.statics) || {};
   this.tree = {};
   this.query = {};
   this.childSchemas = [];
