+++ test/model.test.js
@@ -8487,8 +8487,8 @@ function pick(obj, keys) {
   return newObj;
 }
 
-describe('Check if statics functions that is supplied in schema option is availabe', function() {
-  it('should give an array back rather than undefined m0_0aAutoTyped', function M0_0aModelJS() {
+describe('Check if statics functions that is supplied in schema option is availabe (m0_0a)', function() {
+  it('should give an static function back rather than undefined', function M0_0aModelJS() {
     const testSchema = new mongoose.Schema({}, { statics: { staticFn() { return 'Returned from staticFn'; } } });
     const TestModel = mongoose.model('TestModel', testSchema);
     assert.equal(TestModel.staticFn(), 'Returned from staticFn');
